add conta exceptions

+++ api/src/main/java/br/com/hellobankproject/api/service/conta/IConta.java
@@ -1,6 +1,7 @@
 package br.com.hellobankproject.api.service.conta;
 
 import br.com.hellobankproject.api.dao.ContaDAO;
+import br.com.hellobankproject.api.exception.NotFoundException;
 import br.com.hellobankproject.api.model.Conta;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,34 +12,37 @@ import java.util.ArrayList;
 public class IConta implements IContaService {
     @Autowired
     private ContaDAO dao;
+    private static final String MSG = "Conta nao cadastrada";
+
     @Override
     public Conta criarNovoConta(Conta novo) {
-        if (novo.getTipo() != null && novo.getSaldo() != null) {
-            return dao.save(novo);
-        }
-        return null;
+        return dao.save(novo);
     }
 
     @Override
     public Conta atualizarDadosConta(Conta dados) {
-        if (dados.getId() != null && dados.getTipo() != null && dados.getSaldo() != null) {
+        if (dao.existsById(dados.getId())) {
             return dao.save(dados);
         }
-        return null;
+        throw new NotFoundException(MSG);
     }
 
     @Override
     public ArrayList<Conta> buscarTodasContas() {
-        return (ArrayList<Conta>)dao.findAll();
+        return (ArrayList<Conta>) dao.findAll();
     }
 
     @Override
     public Conta buscarPeloIdConta(Integer id) {
-        return dao.findById(id).orElse(null);
+        return dao.findById(id).orElseThrow(() -> new NotFoundException(MSG));
     }
 
     @Override
     public void excluirConta(Integer id) {
-        dao.deleteById(id);
+        if (dao.existsById(id)) {
+            dao.deleteById(id);
+        } else {
+            throw new NotFoundException(MSG);
+        }
     }
 }
